+++ mayan/apps/appearance/settings.py
@@ -8,6 +8,9 @@ from .literals import DEFAULT_MAXIMUM_TITLE_LENGTH
 
 namespace = Namespace(name='appearance', label=_('Appearance'))
 setting_max_title_length = namespace.add_setting(
-    global_name='APPEARANCE_MAXIMUM_TITLE_LENGTH',
-    default=DEFAULT_MAXIMUM_TITLE_LENGTH
+    default=DEFAULT_MAXIMUM_TITLE_LENGTH,
+    global_name='APPEARANCE_MAXIMUM_TITLE_LENGTH', help_text=_(
+        'Maximum number of characters that will be displayed as the view '
+        'title.'
+    )
 )