+++ mayan/apps/lock_manager/settings.py
@@ -11,10 +11,16 @@ namespace = Namespace(name='lock_manager', label=_('Lock manager'))
 
 setting_backend = namespace.add_setting(
     default=DEFAULT_BACKEND,
-    global_name='LOCK_MANAGER_DEFAULT_BACKEND',
+    global_name='LOCK_MANAGER_BACKEND', help_text=_(
+        'Path to the class to use when to request and release '
+        'resource locks.'
+    )
 )
 
 setting_default_lock_timeout = namespace.add_setting(
     default=DEFAULT_LOCK_TIMEOUT_VALUE,
-    global_name='LOCK_MANAGER_DEFAULT_LOCK_TIMEOUT',
+    global_name='LOCK_MANAGER_DEFAULT_LOCK_TIMEOUT', help_text=_(
+        'Default amount of time in seconds after which a resource '
+        'lock will be automatically released.'
+    )
 )